Refactor/the node job part 18 (#1037)
* node->peer rename

* add exception log

* removing duplicate message read from network server

handshake is now handled in NetworkPeer

* node's version has wrong AddressFrom

* conflict resolution

* conflict + node->peer

* conflict resolution

* conflict resolution

* conflict resolution nodes->peers

* revert wrong version

* separate event calling from processing

* conflict resolution

* refactor message producer and listener to be able to correctly dispose it

* redesign peer disconnect and dispose mechanism

* remove unused code

* disconnect all network peers

* distinguish between using disconnect and dispose methods for better readability

+ comment on NetworkPeer class

* make message producer proxy in network peer

* fix log

* fix tests

* always have disconnect/dispose reason

* fix unit tests

* prevent deadlock by changing the state after client is disposed and event set

* call cancellation callback in separate context of the cancellation caller

* improve log

* prevent firing extra events when going offline

* beat final deadlock boss with local variable for state to be set later

* remove empty line

* this again ... can we put it to gitignore?

* remove unused usings

* final polish - separate vresion message handling

* comment back

* fix xmldoc

* conflict resolution

* Part 18 starts here

otherwise this commit is useless

* construction of peer clients is moved close to construction of peer connections

this is the first step in the process of merging peer client class with peer connection

* move NetworkPeerConnection and NetworkPeerClient to single file

* move NetworkPeerConnection and NetworkPeerClient to one file

* move creation of network peer clients from factory to connection class

we temporarily break client identifiers with this

* peer connection objects are now created using factory

client ID functionality is restored

* no longer we need client in message

* merge NetworkPeerClient with NetworkPeerConnection

* hiding methods that are no longer needed to be public

* moving properties and methods to better order in the merged class

* simplify shutdown sequence

* rename dispose peer client to disconnect

* remove shutdown completion event, use task completion instead

* seems like a bug to add listener in remove method

* put the disposed flag check to dispose method

* comment fix

* finally achieve the correct shutdown sequence

* no need to check if cancellation was done or not, it is done in Cancel call

* modify comment

no longer the deadlock should be possible, but we still prefer separated context

* language and code style fix

* removing cancellation registration callback

i found out that calling CancellationTokenSource.Cancel executes in the context of the calling thread the following two things in this particular order:

1) awakes waiting tasks that are waiting on its cancellation token
2) calls registration callback

because 2) is called after 1) we did have a deadlock in the shutdown procedure as tasks waiting for the cancellation token executed Dispose method

so this commit is replacing the registration with manual call of the shutdown method inside a separated context + ensuring the shutdown is called only once

* conflict resolution
