+++ src/Stratis.Bitcoin/P2P/Protocol/IncomingMessage.cs
@@ -8,7 +8,6 @@ namespace Stratis.Bitcoin.P2P.Protocol
     public class IncomingMessage
     {
         public Message Message { get; set; }
-        internal NetworkPeerClient Client { get; set; }
         public NetworkPeer NetworkPeer { get; set; }
         public long Length { get; set; }
